UAST: move util for converting source to light element
